Refactor metrics trackers to their own file
